+++ boards/stm32wb55xx_nucleo/board.h
@@ -13,7 +13,7 @@ extern whal_Flash g_whalFlash;
 extern whal_Rng g_whalRng;
 extern whal_Crypto g_whalCrypto;
 
-extern volatile size_t g_tick;
+extern volatile uint32_t g_tick;
 
 #define BOARD_LED_PIN         0
 #define BOARD_FLASH_TEST_ADDR 0x08080000
